support config file
Enable the support of using a config file json for the pentoo-installer. No idea how the file is configured but you can reference it and pull it into the VM now rather than fully relying on boot_commands.

I just pulled the json from the pentoo-installer repo, not sure how to configure that file but it's possible now.