post-processors add
Added the post-processors block for uploading the resulting OVA to AWS and converting it to an AMI. I don't know if it works because I don't understand how AWS perms work and I'm tired but it should work if you have the vmimport role and it is configured properly.

Updated README to include new use and some explanation around it.